Güvenlik Önlemleri
Admin Areada rol bazlı yetkilendirmeye göre güvenlik önlemleri alındı , eğer kullanıcı rolü admin değil ise bu areda erişim reddedildi hatası yerine 404 notfound hatası alıcak böylece URL lerden haberdar olması engellendi
User areada ise mail doğrulama işlemine göre işlemlere erişmesi engellendi ve mail doğrulama sayfasına yönlendirildi

+++ BugsByteLibrary/Areas/Admin/Controllers/AdminCodeBankBlogController.cs
@@ -1,10 +1,12 @@
 ﻿using Core.Layer.IService;
 using Core.Layer.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Service.Layer.FluentValidations;
 
 namespace BugsByteLibrary.Areas.Admin.Controllers
 {
+    [Authorize]
     [Area("Admin")]
     public class AdminCodeBankBlogController : Controller
     {
@@ -20,6 +22,12 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
         [HttpGet]
         public async Task<IActionResult> Index()
         {
+
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
+
             var values = await _blogservice.GetAllBlogAsync();
 
             return View(values);
@@ -32,6 +40,10 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
         [HttpGet]
         public async Task<IActionResult> AddBlog()
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
 
             ViewBag.Categories = await _categoryService.GetAllCategoryAsync();//categorileri bir checkbox nesnesine atayabilmek için
 
@@ -43,7 +55,10 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
         [HttpPost]
         public async Task<IActionResult> AddBlog(Blog blog, List<int> SelectedCategoryIds)
         {
-
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
 
             if (blog.Image != null)
 
@@ -100,6 +115,10 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
         [HttpGet]
         public async Task<IActionResult> UpdateBlog(int id)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
 
             var blog = await _blogservice.GetBlogByIdAsync(id);
 
@@ -112,7 +131,10 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
         [HttpPost]
         public async Task<IActionResult> UpdateBlog(Blog blog, List<int> SelectedCategoryIds)
         {
-
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
 
             blog.UpdateDate = DateTime.Now;
             blog.Status = true;
@@ -151,6 +173,10 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
 
         public async Task<IActionResult> DeleteBlog(int id)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
 
             var blog = await _blogservice.GetBlogByIdAsync(id);
             await _blogservice.DeleteBlog(blog);
@@ -163,6 +189,12 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
 
         public async Task<IActionResult> ChangeStatusFalse(Blog blog, int id)
         {
+
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
+
             blog = await _blogservice.GetBlogByIdAsync(id);
             blog.Status = false;
 
@@ -173,6 +205,12 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
 
         public async Task<IActionResult> ChangeStatusTrue(Blog blog, int id)
         {
+
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
+
             blog = await _blogservice.GetBlogByIdAsync(id);
             blog.Status = true;
