+++ BugsByteLibrary/Areas/Admin/Controllers/AdminHiringController.cs
@@ -1,10 +1,12 @@
 ﻿using Core.Layer.IService;
 using Core.Layer.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Service.Layer.Services;
 
 namespace BugsByteLibrary.Areas.Admin.Controllers
 {
+    [Authorize]
     [Area("Admin")]
     public class AdminHiringController : Controller
     {
@@ -19,6 +21,11 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
 
         public async Task<IActionResult> Index()
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
+
             var hirings = await _hiringService.GetAllHiringAsync();
 
             return View(hirings);
@@ -26,6 +33,12 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
 
         public async Task<IActionResult> GetHiringDetails(string id)
         {
+
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
+
             var hiring = await _hiringService.GetHiringByIdAsync(id);
 
             TempData["hiringId"] = hiring.Id;
@@ -43,6 +56,11 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
         [HttpPost]
         public async Task<IActionResult> AddHiring(Hiring hiring)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
+
             await _hiringService.AddHiringAsync(hiring);
 
             return RedirectToAction(nameof(Index));
@@ -51,6 +69,11 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
         [HttpGet]
         public async Task<IActionResult> UpdateHiring(string id)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
+
             var hiring = await _hiringService.GetHiringByIdAsync(id);
 
             return View(hiring);
@@ -60,7 +83,12 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
         [HttpPost]
         public async Task<IActionResult> UpdateHiring(Hiring hiring )
         {
-            
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
+
+
             hiring.UpdateDate = DateTime.Now;
             hiring.Status = true;
             await _hiringService.UpdateHiringAsync(hiring);
@@ -69,6 +97,12 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
         }
         public async Task<IActionResult> DeleteHiring(string id)
         {
+
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
+
             var hiring = await _hiringService.GetHiringByIdAsync(id);
 
             await _hiringService.DeleteHiringAsync(hiring);
@@ -78,6 +112,12 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
 
         public async Task<IActionResult> ChangeStatusFalse(Hiring hiring, string id)
         {
+
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
+
             hiring = await _hiringService.GetHiringByIdAsync(id);
             hiring.Status = false;
 
@@ -88,6 +128,12 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
 
         public async Task<IActionResult> ChangeStatusTrue(Hiring hiring, string id)
         {
+
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
+
             hiring = await _hiringService.GetHiringByIdAsync(id);
             hiring.Status = true;
 
@@ -98,7 +144,11 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
 
         public async Task<IActionResult> ChangeOpenToWorkStatusFalse(string id)
         {
-           
+
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
 
             var openToWork = await _openToWorkService.GetOpenToWorkByIdAsnc(id);
             openToWork.Status = false;
@@ -113,7 +163,10 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
 
         public async Task<IActionResult> ChangeOpenToWorkStatustrue(string id )
         {
-            
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
 
 
             var openToWork = await _openToWorkService.GetOpenToWorkByIdAsnc(id);
@@ -125,7 +178,12 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
         }
         public IActionResult DownloadPdf(string pdfName)
         {
-            
+
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
+
             var filePath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "cvpdf", pdfName);
