+++ BugsByteLibrary/Areas/Admin/Controllers/BlogCategoriesController.cs
@@ -6,7 +6,7 @@ using System.Security.Cryptography.X509Certificates;
 
 namespace BugsByteLibrary.Areas.Admin.Controllers
 {
-    [Authorize(Roles ="Admin")]
+    [Authorize]
     [Area("Admin")]
     public class BlogCategoriesController : Controller
     {
@@ -21,6 +21,12 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
         public async Task<IActionResult> Index()
         {
 
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
+
+
             var category = await _categoryService.GetAllCategoryAsync();
 
             return View(category);
@@ -30,11 +36,21 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
         [HttpGet]
         public IActionResult AddCategory()
         {
+
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
+
             return View();
         }
         [HttpPost]
         public async Task<IActionResult> AddCategory(Category category)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
 
             await _categoryService.AddCategoryAsync(category);
             return RedirectToAction(nameof(Index));
@@ -43,6 +59,10 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
         [HttpGet]
         public async Task<IActionResult> EditCategory(int id, Category category)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
 
             category = await _categoryService.GetByCategoryIdAsync(id);
             return View(category);
@@ -52,12 +72,22 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
         [HttpPost]
         public async Task<IActionResult> EditCategory(Category category)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
+
             await _categoryService.UpdateCategoryAsync(category);
             return RedirectToAction(nameof(Index));
         }
 
         public async Task<IActionResult> DeleteCategory(Category category, int id)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
+
             category = await _categoryService.GetByCategoryIdAsync(id);
 
             await _categoryService.DeleteCategoryAsync(category);
