+++ BugsByteLibrary/Areas/Admin/Controllers/BookController.cs
@@ -1,9 +1,11 @@
 ﻿using Core.Layer.IService;
 using Core.Layer.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace BugsByteLibrary.Areas.Admin.Controllers
 {
+    [Authorize]
     [Area("Admin")]
     public class BookController : Controller
     {
@@ -16,17 +18,34 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
 
         public async Task<IActionResult> Index()
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
+
+
             return View(await _bookService.GetAllBookAsync());
         }
 
         public IActionResult AddBook()
         {
+
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
+
+
             return View();
         }
 
         [HttpPost]
         public async Task<IActionResult> AddBook(EBook book)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
 
 
             if (book.Pdf != null && book.Pdf.Length > 0)
@@ -48,7 +67,14 @@ namespace BugsByteLibrary.Areas.Admin.Controllers
 
         public async Task<IActionResult> DeleteBook(int id , EBook book) 
         {
-             book = await _bookService.GetBookByIdAsync(id);
+
+            if (!User.IsInRole("Admin"))
+            {
+                return NotFound();
+            }
+
+
+            book = await _bookService.GetBookByIdAsync(id);
 
              _bookService.DeleteBookAsync(book);
